+++ src/main/java/com/example/swcompetitionproject/controller/AuthController.java
@@ -4,8 +4,8 @@ import com.example.swcompetitionproject.authentication.AuthenticationExtractor;
 import com.example.swcompetitionproject.authentication.JwtEncoder;
 import com.example.swcompetitionproject.dto.request.auth.LoginDto;
 import com.example.swcompetitionproject.dto.response.auth.LoginResponseDto;
-import com.example.swcompetitionproject.dto.response.auth.ResponseDto;
-import com.example.swcompetitionproject.dto.response.TokenResponseDto;
+import com.example.swcompetitionproject.dto.response.ResponseDto;
+import com.example.swcompetitionproject.dto.response.auth.TokenResponseDto;
 import com.example.swcompetitionproject.service.LoginService;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
