+++ src/main/java/com/example/swcompetitionproject/dto/response/auth/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.swcompetitionproject.dto.response;
+package com.example.swcompetitionproject.dto.response.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
